tebsarvis-multi-agent/requirements.txt
# Core Dependencies
azure-functions==1.23.0
azure-search-documents>=11.4.0
azure-openai>=1.6.0
azure-cosmos>=4.5.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
openpyxl>=3.1.0
python-dateutil>=2.8.0

# Machine Learning & Embeddings
faiss-cpu>=1.7.4  # or faiss-gpu for GPU support
sentence-transformers>=2.2.0
scikit-learn>=1.3.0

# Async & HTTP
aiohttp>=3.9.0
asyncio-throttle>=1.0.0
httpx>=0.25.0

# Text Processing
nltk>=3.8.0
regex>=2023.10.0
beautifulsoup4>=4.12.0

# Logging & Monitoring
structlog>=23.2.0
rich>=13.7.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
tenacity>=8.2.0
click>=8.1.0

# Development (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.10.0
flake8>=6.1.0
mypy>=1.7.0